Fix the Wi-Fi initialization in case of TizenRT.
JSRemoteTest-DCO-1.0-Signed-off-by: Roland Takacs rtakacs.uszeged@partner.samsung.com
